Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -1,11 +1,11 @@
 #include<stdio.h>
 float num1,num2,product;
 void multiply(){
-    product=num1*num2;
-    printf("Product: %d",product)
+    product=(float)num1*num2;
+    printf("Product: %f",product)
 }
 int main(){
-    scanf("%d %d",&num1,&num2);
+    scanf("%f %f",&num1,&num2);
     multiply();
     return 0;
 }
